[widgets] Extract ImageCropper and reuse it for production picture
The crop / drag / zoom logic that lived inside ChangeAvatarModal now
lives in a standalone widget consumed by both ChangeAvatarModal (avatar)
and EditProductionModal (production picture).

- Add src/components/widgets/ImageCropper.vue. Props: `shape` ('circle'
  or 'rounded'), `size` (display, default 180px), `outputSize` (canvas,
  default 400px), `outputType` / `outputQuality`. Emits `fileselected`
  with the raw FormData when an image is picked or dropped. Exposes
  `cropToFormData()` and `reset()` via defineExpose so parents can
  request the framed FormData on confirm.
- Slim ChangeAvatarModal down to a thin wrapper that forwards
  `fileselected` and calls `cropToFormData()` from its confirm handler.
- Replace the FileUpload + label block in EditProductionModal's
  picture field with the cropper (shape="rounded"); make
  `runConfirmation` async so the framed FormData is re-emitted via
  `fileselected` before the existing `confirm` payload reaches the
  parent. The existing storeProductionPicture + uploadProductionAvatar
  flow keeps working since it consumes whatever was last stored.
